+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - `ImmutableCollection`, `ImmutableList`, `ImmutableMap`, `ImmutableSet`: added clarifying JavaDoc
 - `MiniMap0`, `MiniSet0`: minor optimisations
 - `pom.xml`: fixed repository reference
+- `ImmutableCollection`, `ImmutableList`, `ImmutableSet`: optimised constructors (removed duplicate length checks)
 
 ## [2.3] - 2023-01-08
 ### Changed